requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
distro==1.9.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
langcodes==3.3.0
MarkupSafe==2.1.5
murmurhash==1.0.10
numpy==1.26.4
openai==1.13.3
packaging==23.2
passlib==1.7.4
preshed==3.0.9
pyasn1==0.5.1
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
srsly==2.4.8
starlette==0.37.1
thinc==8.2.3
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
